newbie
Hi, how are you doing?

I'm a student currently taking an open source class. I'm completely new to Github so I'm not sure how else to make contact. Part of my grade involves joining and contributing to an open source project and this one sounds rather interesting. I'm not all that skilled at coding but I am an intermediate guitarist so I may be able to help with any related knowledge. Otherwise, if there are other ways I can help, I'd be happy to. As an amateur guitarist, this would be a pretty cool project to be a part of. You can reach me at gns69753@gmail.com

Thanks